requirements.txt
Flask
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
openai
requests
python-dotenv